Provide temporary secondary artifact for no-cp runs
This secondary artifact simply carries a module-info that is meant to
bring onto the module path all JARs that are not actually modules and
should thus be located on the module path.

Whether this artifact should be provided is currently being
investigated, so you should not rely on its existence.

Signed-off-by: TheSilkMiner <thesilkminer@outlook.com>