Add workaround to generate scripts in the proper location
Turns out after extensive analysis with a colleague on Discord and lots
of hours wasted that the `defaultIdentifier` property is completely
ignored when using the K2 compiler (which is what we are using, for
obvious reasons). This also uncovered a 6-year-old issue about being
able to change the package of generated scripts without `package`
declarations that had no answer or workarounds.

Therefore, the only solution to ensure stuff can be properly
modularized, which requires package names, and with the names we want
for debugging requires editing the source code directly and modify
script information with a name that ends only in ".kts" (surprisingly,
the Kotlin compiler doesn't care about the file extension).

This has the unfortunate side effects of messing with the SourceCode
attribute in the generated class file (which now loses the actual script
name) and with line numbers, thus with error messages in case of syntax
errors and whatnot. Regardless, we'll figure these out later on
(probably ASM for the first and... I guess error message manipulation
for the second?).

Signed-off-by: TheSilkMiner <thesilkminer@outlook.com>